Removed things not avaible in newer api versions, changed some of the apps information and compiled a new version using the newest android sdk version 25

+++ src/org/lfx/azilink/net/VpnLink.java
@@ -237,7 +237,7 @@ public class VpnLink extends SocketHandler {
 		}
 		setPingTimer();
 		mOutput.order( ByteOrder.BIG_ENDIAN );
-		mOutput.putShort( (short)(length) );
+		mOutput.putShort(length);
 		mOutput.put( d, 0, length );
 		mOutput.flip();
 		try {
